// web/src/api/models/document.ts
export interface Document {
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'status'?: DocumentStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'vector_index_status'?: DocumentVectorIndexStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'fulltext_index_status'?: DocumentFulltextIndexStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'graph_index_status'?: DocumentGraphIndexStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'summary_index_status'?: DocumentSummaryIndexStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'vision_index_status'?: DocumentVisionIndexStatusEnum;
    /**
     * Vector index last updated time
     * @type {string}
     * @memberof Document
     */
    'vector_index_updated'?: string;
    /**
     * Fulltext index last updated time
     * @type {string}
     * @memberof Document
     */
    'fulltext_index_updated'?: string;
    /**
     * Graph index last updated time
     * @type {string}
     * @memberof Document
     */
    'graph_index_updated'?: string;
    /**
     * Summary index last updated time
     * @type {string}
     * @memberof Document
     */
    'summary_index_updated'?: string;
    /**
     * Vision index last updated time
     * @type {string}
     * @memberof Document
     */
    'vision_index_updated'?: string;
    /**
     * Summary of the document
     * @type {string}
     * @memberof Document
     */
    'summary'?: string;
    /**
     * 
     * @type {number}
     * @memberof Document
     */
    'size'?: number;
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'created'?: string;
    /**
     * 
     * @type {string}
     * @memberof Document
     */
    'updated'?: string;
}

export const DocumentStatusEnum = {
    UPLOADED: 'UPLOADED',
    EXPIRED: 'EXPIRED',
    PENDING: 'PENDING',
    RUNNING: 'RUNNING',
    COMPLETE: 'COMPLETE',
    FAILED: 'FAILED',
    DELETING: 'DELETING',
    DELETED: 'DELETED'
} as const;

export type DocumentStatusEnum = typeof DocumentStatusEnum[keyof typeof DocumentStatusEnum];
export const DocumentVectorIndexStatusEnum = {
    PENDING: 'PENDING',
    CREATING: 'CREATING',
    ACTIVE: 'ACTIVE',
    DELETING: 'DELETING',
    DELETION_IN_PROGRESS: 'DELETION_IN_PROGRESS',
    FAILED: 'FAILED',
    SKIPPED: 'SKIPPED'
} as const;

export type DocumentVectorIndexStatusEnum = typeof DocumentVectorIndexStatusEnum[keyof typeof DocumentVectorIndexStatusEnum];
export const DocumentFulltextIndexStatusEnum = {
    PENDING: 'PENDING',
    CREATING: 'CREATING',
    ACTIVE: 'ACTIVE',
    DELETING: 'DELETING',
    DELETION_IN_PROGRESS: 'DELETION_IN_PROGRESS',
    FAILED: 'FAILED',
    SKIPPED: 'SKIPPED'
} as const;

export type DocumentFulltextIndexStatusEnum = typeof DocumentFulltextIndexStatusEnum[keyof typeof DocumentFulltextIndexStatusEnum];
export const DocumentGraphIndexStatusEnum = {
    PENDING: 'PENDING',
    CREATING: 'CREATING',
    ACTIVE: 'ACTIVE',
    DELETING: 'DELETING',
    DELETION_IN_PROGRESS: 'DELETION_IN_PROGRESS',
    FAILED: 'FAILED',
    SKIPPED: 'SKIPPED'
} as const;

export type DocumentGraphIndexStatusEnum = typeof DocumentGraphIndexStatusEnum[keyof typeof DocumentGraphIndexStatusEnum];
export const DocumentSummaryIndexStatusEnum = {
    PENDING: 'PENDING',
    CREATING: 'CREATING',
    ACTIVE: 'ACTIVE',
    DELETING: 'DELETING',
    DELETION_IN_PROGRESS: 'DELETION_IN_PROGRESS',
    FAILED: 'FAILED',
    SKIPPED: 'SKIPPED'
} as const;

export type DocumentSummaryIndexStatusEnum = typeof DocumentSummaryIndexStatusEnum[keyof typeof DocumentSummaryIndexStatusEnum];
export const DocumentVisionIndexStatusEnum = {
    PENDING: 'PENDING',
    CREATING: 'CREATING',
    ACTIVE: 'ACTIVE',
    DELETING: 'DELETING',
    DELETION_IN_PROGRESS: 'DELETION_IN_PROGRESS',
    FAILED: 'FAILED',
    SKIPPED: 'SKIPPED'
} as const;

export type DocumentVisionIndexStatusEnum = typeof DocumentVisionIndexStatusEnum[keyof typeof DocumentVisionIndexStatusEnum];